# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mitus AI Sports Analytics API
 * OpenAPI spec version: 0.1.0
 */
export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export type PlayerInjuryRisk =
  (typeof PlayerInjuryRisk)[keyof typeof PlayerInjuryRisk];

export const PlayerInjuryRisk = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type PlayerStatus = (typeof PlayerStatus)[keyof typeof PlayerStatus];

export const PlayerStatus = {
  active: "active",
  injured: "injured",
  recovering: "recovering",
} as const;

export interface WearableData {
  heartRate: number;
  heartRateMax: number;
  speed: number;
  distance: number;
  acceleration: number;
  fatigue: number;
  lastUpdated: string;
}

export interface Player {
  id: number;
  name: string;
  number: number;
  position: string;
  age: number;
  nationality: string;
  height: number;
  weight: number;
  muscleMass: number;
  injuryRisk: PlayerInjuryRisk;
  status: PlayerStatus;
  avatarUrl?: string | null;
  wearableData: WearableData;
}

export type MedicalRecordClearanceStatus =
  (typeof MedicalRecordClearanceStatus)[keyof typeof MedicalRecordClearanceStatus];

export const MedicalRecordClearanceStatus = {
  cleared: "cleared",
  restricted: "restricted",
  "not-cleared": "not-cleared",
} as const;

export interface InjuryRecord {
  type: string;
  date: string;
  recovered: boolean;
  notes: string;
}

export interface MedicalRecord {
  playerId: number;
  bloodType: string;
  allergies: string[];
  injuries: InjuryRecord[];
  medications: string[];
  notes: string;
  lastExamDate: string;
  clearanceStatus: MedicalRecordClearanceStatus;
}

export type SessionType = (typeof SessionType)[keyof typeof SessionType];

export const SessionType = {
  match: "match",
  training: "training",
  recovery: "recovery",
} as const;

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus];

export const SessionStatus = {
  scheduled: "scheduled",
  "in-progress": "in-progress",
  completed: "completed",
} as const;

export interface Session {
  id: number;
  type: SessionType;
  date: string;
  duration: number;
  playerIds: number[];
  description: string;
  status: SessionStatus;
}

export interface MovementAnalysis {
  id: number;
  playerId: number;
  sessionId: number;
  timestamp: string;
  movementType: string;
  correct: boolean;
  riskScore: number;
  detectedIssues: string[];
  bodyParts: string[];
  recommendation: string;
  confidence: number;
}

export type PreventionProtocolCategory =
  (typeof PreventionProtocolCategory)[keyof typeof PreventionProtocolCategory];

export const PreventionProtocolCategory = {
  "warm-up": "warm-up",
  stretching: "stretching",
  strength: "strength",
  recovery: "recovery",
  mobility: "mobility",
} as const;

export type PreventionProtocolPriority =
  (typeof PreventionProtocolPriority)[keyof typeof PreventionProtocolPriority];

export const PreventionProtocolPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Exercise {
  name: string;
  sets: number;
  reps: string;
  duration: string;
  notes: string;
}

export interface PreventionProtocol {
  id: number;
  playerId: number;
  title: string;
  category: PreventionProtocolCategory;
  exercises: Exercise[];
  frequency: string;
  priority: PreventionProtocolPriority;
  createdAt: string;
}

export interface DashboardStats {
  totalPlayers: number;
  activePlayers: number;
  injuredPlayers: number;
  highRiskPlayers: number;
  totalSessions: number;
  recentAnalyses: number;
  teamFatigue: number;
  teamAvgHeartRate: number;
}

export interface CreatePlayerInput {
  name: string;
  number: number;
  position: string;
  age: number;
  nationality: string;
  height: number;
  weight: number;
  muscleMass: number;
}

export type UpdatePlayerInputInjuryRisk =
  (typeof UpdatePlayerInputInjuryRisk)[keyof typeof UpdatePlayerInputInjuryRisk];

export const UpdatePlayerInputInjuryRisk = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type UpdatePlayerInputStatus =
  (typeof UpdatePlayerInputStatus)[keyof typeof UpdatePlayerInputStatus];

export const UpdatePlayerInputStatus = {
  active: "active",
  injured: "injured",
  recovering: "recovering",
} as const;

export interface UpdatePlayerInput {
  name?: string;
  number?: number;
  position?: string;
  age?: number;
  nationality?: string;
  height?: number;
  weight?: number;
  muscleMass?: number;
  injuryRisk?: UpdatePlayerInputInjuryRisk;
  status?: UpdatePlayerInputStatus;
}

export type UpdateMedicalInputClearanceStatus =
  (typeof UpdateMedicalInputClearanceStatus)[keyof typeof UpdateMedicalInputClearanceStatus];

export const UpdateMedicalInputClearanceStatus = {
  cleared: "cleared",
  restricted: "restricted",
  "not-cleared": "not-cleared",
} as const;

export interface UpdateMedicalInput {
  bloodType?: string;
  allergies?: string[];
  medications?: string[];
  notes?: string;
  lastExamDate?: string;
  clearanceStatus?: UpdateMedicalInputClearanceStatus;
}

export type CreateSessionInputType =
  (typeof CreateSessionInputType)[keyof typeof CreateSessionInputType];

export const CreateSessionInputType = {
  match: "match",
  training: "training",
  recovery: "recovery",
} as const;

export interface CreateSessionInput {
  type: CreateSessionInputType;
  date: string;
  duration: number;
  playerIds: number[];
  description: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type DeletePlayer200 = {
  message?: string;
};

export type DeleteSession200 = {
  message?: string;
};

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};