# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mitus AI Sports Analytics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Add a new player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlayerInput"
      responses:
        "201":
          description: Player created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get a player by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Not found
    put:
      operationId: updatePlayer
      tags: [players]
      summary: Update a player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlayerInput"
      responses:
        "200":
          description: Player updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    delete:
      operationId: deletePlayer
      tags: [players]
      summary: Delete a player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "404":
          description: Not found
  /players/{id}/medical:
    get:
      operationId: getPlayerMedical
      tags: [players]
      summary: Get medical record for a player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medical record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
    put:
      operationId: updatePlayerMedical
      tags: [players]
      summary: Update medical record for a player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMedicalInput"
      responses:
        "200":
          description: Medical record updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
...
components:
...
  schemas:
...
    Player:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        number:
          type: integer
        position:
          type: string
        age:
          type: integer
        nationality:
          type: string
        height:
          type: number
        weight:
          type: number
        muscleMass:
          type: number
        injuryRisk:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [active, injured, recovering]
        avatarUrl:
          type: string
          nullable: true
        wearableData:
          $ref: "#/components/schemas/WearableData"
      required:
        - id
        - name
        - number
        - position
        - age
        - nationality
        - height
        - weight
        - muscleMass
        - injuryRisk
        - status
        - wearableData
    WearableData:
      type: object
      properties:
        heartRate:
          type: integer
        heartRateMax:
          type: integer
        speed:
          type: number
        distance:
          type: number
        acceleration:
          type: number
        fatigue:
          type: integer
        lastUpdated:
          type: string
      required:
        - heartRate
        - heartRateMax
        - speed
        - distance
        - acceleration
        - fatigue
        - lastUpdated
    MedicalRecord:
      type: object
      properties:
        playerId:
          type: integer
        bloodType:
          type: string
        allergies:
          type: array
          items:
            type: string
        injuries:
          type: array
          items:
            $ref: "#/components/schemas/InjuryRecord"
        medications:
          type: array
          items:
            type: string
        notes:
          type: string
        lastExamDate:
          type: string
        clearanceStatus:
          type: string
          enum: [cleared, restricted, not-cleared]
      required:
        - playerId
        - bloodType
        - allergies
        - injuries
        - medications
        - notes
        - lastExamDate
        - clearanceStatus
    InjuryRecord:
      type: object
      properties:
        type:
          type: string
        date:
          type: string
        recovered:
          type: boolean
        notes:
          type: string
      required:
        - type
        - date
        - recovered
        - notes
...
    CreatePlayerInput:
      type: object
      properties:
        name:
          type: string
        number:
          type: integer
        position:
          type: string
        age:
          type: integer
        nationality:
          type: string
        height:
          type: number
        weight:
          type: number
        muscleMass:
          type: number
      required:
        - name
        - number
        - position
        - age
        - nationality
        - height
        - weight
        - muscleMass
    UpdatePlayerInput:
      type: object
      properties:
        name:
          type: string
        number:
          type: integer
        position:
          type: string
        age:
          type: integer
        nationality:
          type: string
        height:
          type: number
        weight:
          type: number
        muscleMass:
          type: number
        injuryRisk:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [active, injured, recovering]
    UpdateMedicalInput:
      type: object
      properties:
        bloodType:
          type: string
        allergies:
          type: array
          items:
            type: string
        medications:
          type: array
          items:
            type: string
        notes:
          type: string
        lastExamDate:
          type: string
        clearanceStatus:
          type: string
          enum: [cleared, restricted, not-cleared]
...